+++ app/app.js
@@ -3,7 +3,7 @@
 //.hide unatharized view
 
 const authEvents = require('./auth/events')
-const ui = require('./auth/ui')
+const gameEvents = require('./game/events')
 
 // use require without a reference to ensure a file is bundled
 // require('./example')
@@ -11,11 +11,13 @@ const ui = require('./auth/ui')
 $(() => {
   $("#signed_in_user").hide();
   $('#game-board').hide()
+  
   $('#sign-up').on('submit', authEvents.onSignUp)
   $('#sign-in').on('submit', authEvents.onSignIn)
   $('#sign-out').on('click', authEvents.onSignOut)
-  $('#new-game').on('click', authEvents.onNewGame)
-  ui.accessAllCells().on('click', authEvents.onPlay)
+  
+  $('#new-game').on('click', gameEvents.onNewGame)
+  // ui.accessAllCells().on('click', authEvents.onPlay)
   // let currentPlayer = '✕'
 
   // // // Our box click event handler
