+++ app/styles/index.scss
@@ -126,7 +126,7 @@ form{
   // When you hover over a box
   &:hover {
     // Set the background color to a light blue
-    background-color: LightSkyBlue;
+    background-color: rgba(0, 0, 0, 0.5);
   }
 }
